fixed: building components
