+++ next.config.mjs
@@ -1,4 +1,10 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {};
+const nextConfig = {
+    output: 'export',
+    basePath: process.env.BASEPATH,
+    images: {
+        unoptimized: true,
+    },
+};
 
 export default nextConfig;
